Fix tellraw, bossbar

+++ src/main/java/fr/olympa/olympacreatif/commandblocks/CbBossBar.java
@@ -1,8 +1,5 @@
 package fr.olympa.olympacreatif.commandblocks;
 
-import org.bukkit.Bukkit;
-import org.bukkit.boss.BarColor;
-import org.bukkit.boss.BarStyle;
 import org.bukkit.boss.BossBar;
 
 import fr.olympa.olympacreatif.OlympaCreatifMain;
@@ -10,8 +7,8 @@ import fr.olympa.olympacreatif.OlympaCreatifMain;
 public class CbBossBar {
 
 	private BossBar bar;
-	private int value = 100;
-	private int max = 100;
+	private double value = 100;
+	private double max = 100;
 	
 	public CbBossBar(OlympaCreatifMain plugin, BossBar bar) {
 		this.bar = bar;
@@ -34,6 +31,6 @@ public class CbBossBar {
 	}
 
 	public int getMax() {
-		return max;
+		return (int) max;
 	}
 }
